requirements.txt
# requirements.txt

transformers
torch
Pillow
accelerate
faiss-cpu
numpy
gradio
timm
einops